requirements.txt
keyboard==0.13.5
pyperclip==1.8.2
PyAutoGUI==0.9.54
google-genai==1.21
python-dotenv==1.1
